add save to db and get from db

+++ lib/Backend/DB/DAOs/Mappings/mapping_task_date.dart
@@ -17,7 +17,9 @@ class MappingTaskDate {
       ),
       doneDate:
           rawData[TTaskDate.doneDate] != null
-              ? DateTime.fromMillisecondsSinceEpoch(rawData[TTaskDate.doneDate])
+              ? DateTime.fromMillisecondsSinceEpoch(
+                rawData[TTaskDate.doneDate] as int,
+              )
               : null,
       doneBy: doneBy.isNotEmpty ? doneBy : null,
       task: task,
